refactor(client): Extract `LessonDisplay` & `ParticipantListList` into components
